Fix tests in CI by allowing loading posts by id
It seems that when running the tests through ember-try, there's
something that changes causing the acceptance tests to attempt to load
posts by id. My guess is that it has something to do with
`shouldReloadRecord`, but I'm really not sure why that showed up with
ember-try and not when I run the tests normally.